Django/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.21.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
argcomplete==3.5.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.1.3
arxiv-mcp-server==0.2.6
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
coverage==7.6.12
Crawl4AI==0.5.0.post4
cryptography==44.0.0
cssselect==1.3.0
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
Django==5.1.7
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
executing==2.2.0
fake-http-header==0.3.5
fake-useragent==2.1.0
fastapi==0.115.11
fastjsonschema==2.21.1
faust-cchardet==2.1.19
feedparser==6.0.11
ffmpeg==1.4
filelock==3.18.0
Flask==3.1.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
humanize==4.12.1
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.32.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jwcrypto==1.5.6
langchain==0.3.15
langchain-core==0.3.31
langchain-text-splitters==0.3.5
langdetect==1.0.9
langsmith==0.3.1
litellm==1.63.11
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mcp==1.2.0
mcp-server-time==0.6.2
mdurl==0.1.2
mistune==3.1.1
mock==5.1.0
multidict==6.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
nmake==1.6.0
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.2.2
oauthlib==3.2.2
openai==1.66.3
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pefile==2023.2.7
pillow==10.4.0
pipx==1.7.1
platformdirs==4.3.6
playwright==1.50.0
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.0
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.19.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.0
PyJWT==2.9.0
PyMuPDF==1.25.2
pymupdf4llm==0.0.17
pyOpenSSL==25.0.0
pypdfium2==4.30.1
pyperclip==1.9.0
pytesseract==0.3.13
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-django==4.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-multipart==0.0.20
pytube==11.0.2
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.1
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
Send2Trash==1.8.3
setuptools==75.8.0
sgmllib3k==1.0.0
shellingham==1.5.4
signtool==4.0.1
sigtools==4.0.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
SQLAlchemy==2.0.37
sqlparse==0.5.3
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.45.2
tenacity==9.0.0
terminado==0.18.1
tf-playwright-stealth==1.1.2
tiktoken==0.9.0
tinycss2==1.4.0
tk==0.1.0
tokenizers==0.21.1
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-certifi==2021.10.8.3
types-python-dateutil==2.9.0.20241206
types-toml==0.10.8.20240310
typing_extensions==4.12.2
tzdata==2025.1
unidiff==0.7.5
uri-template==1.3.0
urllib3==2.3.0
userpath==1.9.2
uvicorn==0.34.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.18.3
yt-dlp-better-download==1.0.0
yt-download==0.1.1
zipp==3.21.0
zstandard==0.23.0
